# storage/api-docs/api-docs.json
{
    "swagger": "2.0",
    "info": {
        "title": "Digital Registry Api",
        "description": "",
        "termsOfService": "",
        "contact": {
            "email": "contact@mysite.com"
        },
        "version": "1.0.0"
    },
    "host": "api.digital-registry.dk",
    "basePath": "/",
    "schemes": [
        "http",
        "https"
    ],
    "paths": {
        "/contact/{identifier}": {
            "get": {
                "tags": [
                    "contact"
                ],
                "summary": "Get contact by identifier, either CPR or CVR",
                "description": "Return contact.",
                "operationId": "contact",
                "produces": [
                    "application/json"
                ],
                "responses": {
                    "200": {
                        "description": "",
                        "schema": {
                            "type": "array",
                            "items": {
                                "$ref": "#/definitions/Contact"
                            }
                        }
                    },
                    "404": {
                        "description": "Invalid tag value"
                    }
                },
                "security": [
                    {
                        "nemId": [
                            "write:contact",
                            "read:contact"
                        ],
                        "basic-auth": [
                            "read:contact"
                        ]
                    }
                ]
            }
        },
...
        "/contact/{identifier}/medias": {
            "get": {
                "tags": [
                    "contact:media"
                ],
                "summary": "Get contact medias",
                "description": "Get contact medias.",
                "operationId": "contact-medias",
                "produces": [
                    "application/json"
                ],
                "responses": {
                    "201": {
                        "description": "successful operation",
                        "schema": {
                            "type": "array",
                            "items": {
                                "$ref": "#/definitions/Contact"
                            }
                        }
                    },
                    "401": {
                        "description": "Unauthorized"
                    }
                },
                "security": [
                    {
                        "nemId": [
                            "write:contact",
                            "read:contact"
                        ],
                        "basic-auth": [
                            "read:contact"
                        ]
                    }
                ]
            },
            "post": {
                "tags": [
                    "contact:media"
                ],
                "summary": "Create a contact media",
                "description": "Create contact media.",
                "operationId": "contact-medias-create",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "responses": {
                    "201": {
                        "description": "successful operation",
                        "schema": {
                            "type": "array",
                            "items": {
                                "$ref": "#/definitions/Contact"
                            }
                        }
                    },
                    "401": {
                        "description": "Unauthorized"
                    }
                },
                "security": [
                    {
                        "nemId": [
                            "write:contact",
                            "read:contact"
                        ],
                        "basic-auth": [
                            "read:contact"
                        ]
                    }
                ]
            }
        },
        "/contact/{identifier}/medias/{media_id}": {
            "delete": {
                "tags": [
                    "contact:media"
                ],
                "summary": "Delete a contact media",
                "description": "Delete contact media.",
                "operationId": "contact-medias-delete",
                "responses": {
                    "204": {
                        "description": ""
                    },
                    "401": {
                        "description": "Unauthorized"
                    }
                },
                "security": [
                    {
                        "nemId": [
                            "write:contact",
                            "read:contact"
                        ],
                        "basic-auth": [
                            "read:contact"
                        ]
                    }
                ]
            }
        },
...
        "/contact/{identifier}/media-subscriptions/{media_subscription_id}": {
            "delete": {
                "tags": [
                    "contact:media-subscription"
                ],
                "summary": "Delete a contact media subscriptions",
                "description": "Delete contact media.",
                "operationId": "contact-medias-delete",
                "responses": {
                    "204": {
                        "description": ""
                    },
                    "401": {
                        "description": "Unauthorized"
                    }
                },
                "security": [
                    {
                        "nemId": [
                            "write:contact",
                            "read:contact"
                        ],
                        "basic-auth": [
                            "read:contact"
                        ]
                    }
                ]
            }
        },
        "/contact/{identifier}/subscriptions": {
            "get": {
                "tags": [
                    "contact:subscription"
                ],
                "summary": "Get contact subscriptions",
                "description": "Get contact subscriptions.",
                "operationId": "contact-subscription",
                "produces": [
                    "application/json"
                ],
                "responses": {
                    "200": {
                        "description": "",
                        "schema": {
                            "type": "array",
                            "items": {
                                "$ref": "#/definitions/Subscription"
                            }
                        }
                    },
                    "401": {
                        "description": "Unauthorized"
                    }
                },
                "security": [
                    {
                        "nemId": [
                            "write:contact",
                            "read:contact"
                        ],
                        "basic-auth": [
                            "read:contact"
                        ]
                    }
                ]
            },
            "post": {
                "tags": [
                    "contact:subscription"
                ],
                "summary": "Create a contact subscription",
                "description": "Create contact subscription.",
                "operationId": "contact-subscription-create",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "responses": {
                    "201": {
                        "description": "",
                        "schema": {
                            "type": "array",
                            "items": {
                                "$ref": "#/definitions/Subscription"
                            }
                        }
                    },
                    "401": {
                        "description": "Unauthorized"
                    }
                },
                "security": [
                    {
                        "nemId": [
                            "write:contact",
                            "read:contact"
                        ],
                        "basic-auth": [
                            "read:contact"
                        ]
                    }
                ]
            }
        },
        "/contact/{identifier}/subscriptions/{subscription_id}": {
            "get": {
                "tags": [
                    "contact:subscription"
                ],
                "summary": "Get contact subscriptions",
                "description": "Get contact subscriptions.",
                "operationId": "contact-subscription-get",
                "produces": [
                    "application/json"
                ],
                "responses": {
                    "200": {
                        "description": "",
                        "schema": {
                            "type": "array",
                            "items": {
                                "$ref": "#/definitions/Subscription"
                            }
                        }
                    },
                    "401": {
                        "description": "Unauthorized"
                    }
                },
                "security": [
                    {
                        "nemId": [
                            "write:contact",
                            "read:contact"
                        ],
                        "basic-auth": [
                            "read:contact"
                        ]
                    }
                ]
            },
            "put": {
                "tags": [
                    "contact:subscription"
                ],
                "summary": "Update a contact subscription",
                "description": "Create contact subscription.",
                "operationId": "contact-subscription-update",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "responses": {
                    "201": {
                        "description": "",
                        "schema": {
                            "type": "array",
                            "items": {
                                "$ref": "#/definitions/Subscription"
                            }
                        }
                    },
                    "401": {
                        "description": "Unauthorized"
                    }
                },
                "security": [
                    {
                        "nemId": [
                            "write:contact",
                            "read:contact"
                        ],
                        "basic-auth": [
                            "read:contact"
                        ]
                    }
                ]
            },
            "delete": {
                "tags": [
                    "contact:subscription"
                ],
                "summary": "Delete a contact subscription",
                "description": "Delete contact subscription.",
                "operationId": "contact-medias-delete",
                "responses": {
                    "204": {
                        "description": ""
                    },
                    "401": {
                        "description": "Unauthorized"
                    }
                },
                "security": [
                    {
                        "nemId": [
                            "write:contact",
                            "read:contact"
                        ],
                        "basic-auth": [
                            "read:contact"
                        ]
                    }
                ]
            }
        },
        "/contact/{identifier}/subscriptions/{subscription_id}/accept-terms": {
            "put": {
                "tags": [
                    "contact:subscription"
                ],
                "summary": "Accept subscription terms.",
                "description": "Accept subscription terms.",
                "operationId": "contact-subscription-accept-terms",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "responses": {
                    "201": {
                        "description": "",
                        "schema": {
                            "type": "array",
                            "items": {
                                "$ref": "#/definitions/Subscription"
                            }
                        }
                    },
                    "401": {
                        "description": "Unauthorized"
                    }
                },
                "security": [
                    {
                        "nemId": [
                            "write:contact",
                            "read:contact"
                        ],
                        "basic-auth": [
                            "read:contact"
                        ]
                    }
                ]
            }
        },
...
    },
    "definitions": {
...
        "Contact": {},
...
        "Subscription": {},
...
    }
}